requirements.txt
requests==2.31.0
pandas==2.2.1
plotly==5.19.0
python-dotenv==1.0.1
schedule==1.2.1